+++ src/api/services/index.ts
@@ -3,6 +3,7 @@
 *   Copyright © 2023 NatML Inc. All Rights Reserved.
 */
 
+export * from "./environment"
 export * from "./prediction"
 export * from "./predictor"
 export * from "./storage"