+++ src/api/types/environment.ts
@@ -1,0 +1,18 @@
+/*
+*   Function
+*   Copyright © 2023 NatML Inc. All Rights Reserved.
+*/
+
+/**
+ * Predictor environment variable.
+ */
+export interface EnvironmentVariable {
+    /**
+     * Variable name.
+     */
+    name: string;
+    /**
+     * Variable value.
+     */
+    value: string;
+}